requirements.txt
python-dotenv==1.0.0
discord.py==2.3.0
langchain==0.1.0
openai==1.0.0
requests==2.31.0
sentence-transformers
numpy